WW-4043 Moves TestUtils into junit-plugin
