+++ plugins/json/src/test/java/org/apache/struts2/json/JSONPopulatorTest.java
@@ -27,6 +27,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import junit.framework.TestCase;
+import org.apache.struts2.util.TestUtils;
 
 public class JSONPopulatorTest extends TestCase {
